Followed SOLID Principles
